+++ src/cache/inmemory/inMemoryCache.ts
@@ -6,15 +6,16 @@ import { dep, wrap } from 'optimism';
 
 import { ApolloCache, Transaction } from '../core/cache';
 import { Cache } from '../core/types/Cache';
+import { Modifier, Modifiers } from '../core/types/common';
 import { addTypenameToDocument } from '../../utilities/graphql/transform';
+import { StoreObject }  from '../../utilities/graphql/storeUtils';
 import {
   ApolloReducerConfig,
   NormalizedCacheObject,
-  StoreObject,
 } from './types';
 import { StoreReader } from './readFromStore';
 import { StoreWriter } from './writeToStore';
-import { EntityStore, supportsResultCaching, Modifiers, Modifier } from './entityStore';
+import { EntityStore, supportsResultCaching } from './entityStore';
 import {
   defaultDataIdFromObject,
   PossibleTypesMap,
