+++ src/core/watchQueryOptions.ts
@@ -1,7 +1,7 @@
 import { DocumentNode, ExecutionResult } from 'graphql';
 
+import { ApolloCache } from '../cache/core/cache';
 import { FetchResult } from '../link/core/types';
-import { DataProxy } from '../cache/core/types/DataProxy';
 import { MutationQueryReducersMap } from './types';
 import { PureQueryOptions, OperationVariables } from './types';
 
@@ -218,9 +218,9 @@ export interface MutationBaseOptions<
   awaitRefetchQueries?: boolean;
 
   /**
-   * A function which provides a {@link DataProxy} and the result of the
-   * mutation to allow the user to update the store based on the results of the
-   * mutation.
+   * A function which provides an {@link ApolloCache} instance, and the result
+   * of the mutation, to allow the user to update the store based on the
+   * results of the mutation.
    *
    * This function will be called twice over the lifecycle of a mutation. Once
    * at the very beginning if an `optimisticResponse` was provided. The writes
@@ -228,11 +228,6 @@ export interface MutationBaseOptions<
    * this function is called which is when the mutation has succesfully
    * resolved. At that point `update` will be called with the *actual* mutation
    * result and those writes will not be rolled back.
-   *
-   * The reason a {@link DataProxy} is provided instead of the user calling the
-   * methods directly on {@link ApolloClient} is that all of the writes are
-   * batched together at the end of the update, and it allows for writes
-   * generated by optimistic data to be rolled back.
    *
    * Note that since this function is intended to be used to update the
    * store, it cannot be used with a `no-cache` fetch policy. If you're
@@ -284,6 +279,6 @@ export interface MutationOptions<
 
 // Add a level of indirection for `typedoc`.
 export type MutationUpdaterFn<T = { [key: string]: any }> = (
-  proxy: DataProxy,
+  cache: ApolloCache<T>,
   mutationResult: FetchResult<T>,
 ) => void;
